refactor(admin): 관리자 페이지 api 코드 타입 분리
