+++ src/features/admin/apis/index.ts
@@ -1,2 +1,4 @@
 export * from './item.api';
 export * from './rental-check.api';
+export * from './item.type';
+export * from './rental-check.type';
