+++ src/features/admin/apis/rental-check.type.ts
@@ -1,0 +1,3 @@
+export interface RentalCheckRequest {
+  itemId: string;
+}
